// src/constants/interface.tsx
export default interface controlStyleProps {
    readonly x?: number
    readonly y?: number
    readonly width?: number
    readonly height?: number
    readonly fontFamily?: string
    readonly fontSize?: number
    readonly color?: string
    readonly bold?: boolean
    readonly italics?: boolean
    readonly borderOrNot?: boolean
    readonly backgroundColor?: string
    readonly borderType?: string
    readonly borderColor?: string
    readonly borderWidth?: number
    readonly password?: boolean
    readonly placeHolder?: string
    readonly tooltip?: string
    readonly align?: string
    readonly verticalAlign?: string
    readonly visible?: boolean
    readonly enable?: boolean
    readonly underline?: boolean
    readonly isGridChild?: boolean
    readonly headerHeight?: number
    readonly rowHeight?: number
}